+++ cardinal/ecs/options.go
@@ -2,18 +2,18 @@ package ecs
 
 import (
 	"os"
+	"pkg.world.dev/world-engine/cardinal/shard/adapter"
 
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"pkg.world.dev/world-engine/cardinal/ecs/receipt"
 	"pkg.world.dev/world-engine/cardinal/ecs/store"
 	"pkg.world.dev/world-engine/cardinal/events"
-	"pkg.world.dev/world-engine/cardinal/shard"
 )
 
 type Option func(e *Engine)
 
-func WithAdapter(adapter shard.Adapter) Option {
+func WithAdapter(adapter adapter.Adapter) Option {
 	return func(e *Engine) {
 		e.chain = adapter
 	}
