+++ cardinal/shard/adapter/option.go
@@ -1,4 +1,4 @@
-package shard
+package adapter
 
 import "github.com/rotisserie/eris"
