+++ cardinal/world.go
@@ -7,6 +7,8 @@ import (
 	"os"
 	"os/signal"
 	"path/filepath"
+	"pkg.world.dev/world-engine/cardinal/shard/adapter"
+	"pkg.world.dev/world-engine/cardinal/shard/evm"
 	"reflect"
 	"runtime"
 	"strings"
@@ -22,10 +24,8 @@ import (
 	"pkg.world.dev/world-engine/cardinal/ecs/receipt"
 	"pkg.world.dev/world-engine/cardinal/ecs/storage/redis"
 	"pkg.world.dev/world-engine/cardinal/events"
-	"pkg.world.dev/world-engine/cardinal/evm"
 	"pkg.world.dev/world-engine/cardinal/gamestage"
 	"pkg.world.dev/world-engine/cardinal/server"
-	"pkg.world.dev/world-engine/cardinal/shard"
 	"pkg.world.dev/world-engine/cardinal/statsd"
 	"pkg.world.dev/world-engine/cardinal/types/component"
 	"pkg.world.dev/world-engine/cardinal/types/entity"
@@ -184,7 +184,7 @@ func applyProductionOptions(
 		return eris.New("must supply BASE_SHARD_SEQUENCER_ADDRESS and BASE_SHARD_QUERY_ADDRESS for production " +
 			"mode Cardinal worlds")
 	}
-	adapter, err := shard.NewAdapter(shard.AdapterConfig{
+	adapter, err := adapter.NewAdapter(adapter.AdapterConfig{
 		ShardSequencerAddr: cfg.BaseShardSequencerAddress,
 		EVMBaseShardAddr:   cfg.BaseShardQueryAddress,
 	})
